# api-contract-v1.yaml
openapi: 3.1.0
info:
  title: Disaster Early-Warning API
  version: 0.5.0
  description: >
    Contract v1. Region-aware by design; US is the only enabled region in v1.

    CHANGELOG
    0.5.0 — Reports are write-first (fusion runs after the 202):
      ~ ReportAccepted: event_id and tier are now optional (set once fusion has run);
        + report_id, + status (pending | fused | failed)
      + GET /reports/{client_event_id} (the reporter's own report status)
      + 409 on POST /reports when the client_event_id belongs to another reporter; 429 when rate limited
    0.4.0 — Additive only, no breaking changes (per-user alert preferences, marine alerts):
      + AlertPreferences schema; User.alert_preferences
      + PUT /me/alert-preferences
      + GET /alerts: hazard_types, min_severity, include_marine (default to the caller's
        saved preferences when signed in)
      + Alert.product, Alert.marine
    0.3.0 — Additive only, no breaking changes (fields the mobile alert card needs):
      + Alert.hazard_type          (drives the card's hazard icon)
      + Alert.location_name        (place shown in the card's metadata line)
      + Alert.corroboration_count  ("Confirmed by N nearby" on corroborated reports)
    0.2.0 — Additive only, no breaking changes:
      + GET /events/{id}/brief        (Ultra situational brief, async)
      + GET /regions/{region}/cascade (Ultra cascade analysis)
      + GET /guidance/manifest        (offline card sync)
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /me/alert-preferences:
    get:
      tags: [auth]
      summary: The caller's alert preferences (defaults if never set)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertPreferences' }
    put:
      tags: [auth]
      summary: Replace the caller's alert preferences
      description: >
        Applies to GET /alerts and to push delivery alike, so what a user sees and what
        notifies them never disagree. An empty hazard_types list is treated as "all".
        Critical alerts bypass these preferences.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlertPreferences' }
      responses:
        '200':
          description: Saved (normalized) preferences
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertPreferences' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /me/devices:
    post:
      tags: [auth]
      summary: Register a push token for this device
      description: >
        push_provider is explicit rather than inferred so non-FCM regions can be
        added later without a contract change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id, push_provider, push_token, platform]
              properties:
                device_id: { type: string, description: Stable client-generated UUID }
                push_provider: { type: string, enum: [fcm, apns] }
                push_token: { type: string }
                platform: { type: string, enum: [ios, android] }
      responses:
        '204': { description: Registered }
  /me/devices/{device_id}:
    delete:
      tags: [auth]
      summary: Stop push to this device (called on sign-out)
      parameters:
        - name: device_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Removed (or was not registered) }
  /me/near-me:
    put:
      tags: [subscriptions]
      summary: Set the area around the phone, for push
      description: >
        Kept by the app in step with the device position and the Home radius, so
        notifications match what Home lists. Stored at ~1 km precision (2 decimal
        places). Pushes for it start at medium severity; critical always gets through.
        Never appears in GET /subscriptions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [center, radius_km]
              properties:
                center: { $ref: '#/components/schemas/GeoPoint' }
                radius_km: { type: number, maximum: 500 }
      responses:
        '204': { description: Saved }
    delete:
      tags: [subscriptions]
      summary: Stop push for the area around the phone
      responses:
        '204': { description: Removed }

  # ---------- reports ----------
...
  /subscriptions:
    get:
      tags: [subscriptions]
      summary: List region subscriptions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Subscription' }
    post:
      tags: [subscriptions]
      summary: Subscribe to an area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [region, center, radius_km]
              properties:
                region: { $ref: '#/components/schemas/RegionCode' }
                label: { type: string, maxLength: 64 }
                center: { $ref: '#/components/schemas/GeoPoint' }
                radius_km: { type: number, maximum: 500 }
                min_severity:
                  type: string
                  enum: [low, medium, high, critical]
                  default: medium
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
  /subscriptions/{subscription_id}:
    delete:
      tags: [subscriptions]
      summary: Remove a subscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }

  # ---------- mesh ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    RegionCode:
      type: string
      description: >
        Enabled regions. v1 ships US only; CN is reserved and rejected at runtime
        until its feed, coordinate-frame and push work lands.
      enum: [US, CN]
    CoordinateFrame:
      type: string
      description: >
        Declared per region so clients never assume WGS-84. v1 is always WGS84.
      enum: [WGS84, GCJ02]
    GeoPoint:
      type: object
      required: [lat, lon]
      properties:
        lat: { type: number, format: double, minimum: -90, maximum: 90 }
        lon: { type: number, format: double, minimum: -180, maximum: 180 }
        accuracy_m: { type: number }
        frame:
          allOf: [{ $ref: '#/components/schemas/CoordinateFrame' }]
          default: WGS84
    HazardType:
      type: string
      enum: [flood, fire, earthquake, storm, landslide, other]
...
    Severity:
      type: string
      enum: [low, medium, high, critical]
...
    AlertPreferences:
      type: object
      description: >
        What a user wants to be alerted about. Defaults: every hazard type, every
        severity, marine off (marine products are ~70% of NWS alerts and only matter
        on the water). Critical alerts are always delivered: these preferences only
        ever filter lower severities, never a critical alert within the user's radius.
      properties:
        hazard_types:
          type: array
          description: Empty or absent means all.
          items: { $ref: '#/components/schemas/HazardType' }
        min_severity:
          allOf: [{ $ref: '#/components/schemas/Severity' }]
          default: low
        include_marine:
          type: boolean
          default: false
...
    Subscription:
      type: object
      required: [subscription_id, region, center, radius_km, min_severity]
      properties:
        subscription_id: { type: string }
        region: { $ref: '#/components/schemas/RegionCode' }
        label: { type: string }
        center: { $ref: '#/components/schemas/GeoPoint' }
        radius_km: { type: number }
        min_severity: { $ref: '#/components/schemas/Severity' }
...